Getting started document
This document is intended to serve as a reference when onboarding new
students and staff: it contains some of the information that is given to
new colleagues. Moreover, it also contains some suggestions for best
practices (e.g., use Rerun for logging) to avoid people reinventing the
wheel and to promote everyone using (roughly) the same stack.